menu_config/requirements.txt
fastapi
uvicorn
pymongo
pytest
loguru
aio_pika